Minor changes
- Make simpleCounter also 32-bit instead of 8, so IDs are essentially unique
- Set state in serverTimeouts function

+++ RF24Ethernet-lwIP/RF24Ethernet/RF24Client.h
@@ -269,7 +269,7 @@ private:
     
     static char incomingData[2][INCOMING_DATA_SIZE];
     static bool serverActive;
-    static uint8_t simpleCounter;
+    static uint32_t simpleCounter;
     
 #endif
